ML/PitWatch_YOLOv8_nano_AI_model/requirements.txt
fastapi
uvicorn
pydantic
ultralytics
onnxruntime
opencv-python-headless
Pillow
